+++ doc/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 Core
 ~~~~
 
+- Move ``prefix`` and ``suffix`` from ``Contact`` to ``PersonContact``
 - Add ``first_name`` and ``last_name`` fields to ``PersonContact``
 - Add bought with relation to ``ProductCrossSellType``
 - Set customer marketing permission while creating order